+++ app/src/main/java/com/fpf/smartscan/data/MediaIdType.kt
@@ -1,0 +1,8 @@
+package com.fpf.smartscan.data
+
+import com.fpf.smartscan.media.MediaType
+
+data class MediaIdType(
+    val id: Long,
+    val type: MediaType
+)