+++ app/src/main/java/com/fpf/smartscan/data/clusters/ClusterCrossRef.kt
@@ -3,14 +3,12 @@ package com.fpf.smartscan.data.clusters
 import androidx.room.Entity
 import androidx.room.ForeignKey
 import androidx.room.Index
-import androidx.room.TypeConverters
-import com.fpf.smartscan.data.MediaTypeConverter
 import com.fpf.smartscan.data.metadata.MediaMetadata
 import com.fpf.smartscan.media.MediaType
 
 @Entity(
     tableName = "media_cluster_crossref",
-    primaryKeys = ["clusterId", "mediaId"],
+    primaryKeys = ["mediaId", "mediaType"],
     foreignKeys = [
         ForeignKey(
             entity = MediaClusterMetadata::class,
@@ -20,15 +18,17 @@ import com.fpf.smartscan.media.MediaType
         ),
         ForeignKey(
             entity = MediaMetadata::class,
-            parentColumns = ["id"],
-            childColumns = ["mediaId"],
+            parentColumns = ["id", "type"],
+            childColumns = ["mediaId", "mediaType"],
             onDelete = ForeignKey.CASCADE
         )
     ],
-    indices = [Index("mediaId")]
+    indices = [
+        Index(value = ["clusterId"])
+    ]
 )
-@TypeConverters(MediaTypeConverter::class)
 data class ClusterCrossRef(
-    val clusterId: Long,
     val mediaId: Long,
+    val mediaType: MediaType,
+    val clusterId: Long
 )