+++ app/src/main/java/com/fpf/smartscan/data/old/images/ImageTagCrossRefRepository.kt
@@ -1,7 +1,0 @@
-package com.fpf.smartscan.data.old.images
-
-class ImageTagCrossRefRepository(
-    private val dao: ImageTagCrossRefDao,
-) {
-    suspend fun getAllCrossRefs(): List<ImageTagCrossRef> = dao.getAll()
-}